Apply snakecase to response outputs as well.
